conversor_multilang_final/requirements.txt
Flask
Pillow
python-docx
pdf2docx
PyMuPDF
docx2pdf
PyPDF2
gunicorn